+++ src/hoc/asyncComponent.jsx
@@ -4,6 +4,7 @@
 // const SomeAsyncComponent = AsyncComponent(e => import('src/xxx/xxx'))
 
 import React, { PureComponent } from 'react'
+import VL from 'src/utils/viewListen'
 
 const AsyncComponent = comp => {
   class Comp extends PureComponent {
@@ -16,6 +17,9 @@ const AsyncComponent = comp => {
     }
 
     componentDidMount() {
+      // 监听页面变化
+      VL.listen(window.location.pathname)
+
       if (this.hasLoadedComponent()) {
         return
       }