+++ src/utils/viewListen.js
@@ -1,0 +1,5 @@
+const VL = {
+  listen: pathName => {}
+}
+
+export default VL